% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{get_mfa_quanti_var}
\alias{get_mfa_quali_var}
\alias{get_mfa_group}
\alias{fviz_mfa_ind_starplot}
\alias{fviz_mfa_group}
\alias{fviz_mfa_quanti_var}
\alias{fviz_mfa_quali_var}
\alias{get_hmfa_quanti_var}
\alias{get_hmfa_quali_var}
\alias{get_hmfa_group}
\alias{fviz_hmfa_quanti_var}
\alias{fviz_hmfa_quali_var}
\alias{fviz_hmfa_ind_starplot}
\alias{fviz_hmfa_group}
\title{Deprecated Functions}
\usage{
get_mfa_quanti_var(res.mfa)

get_mfa_quali_var(res.mfa)

get_mfa_group(res.mfa)

fviz_mfa_ind_starplot(X, ...)

fviz_mfa_group(X, ...)

fviz_mfa_quanti_var(X, ...)

fviz_mfa_quali_var(X, ...)

get_hmfa_quanti_var(res.hmfa)

get_hmfa_quali_var(res.hmfa)

get_hmfa_group(res.hmfa)

fviz_hmfa_quanti_var(X, ...)

fviz_hmfa_quali_var(X, ...)

fviz_hmfa_ind_starplot(X, ...)

fviz_hmfa_group(X, ...)
}
\arguments{
\item{res.mfa}{an object of class MFA [FactoMineR].}

\item{X}{an object of class MFA or HMFA [FactoMineR].}

\item{...}{Other arguments.}

\item{res.hmfa}{an object of class HMFA [FactoMineR].}
}
\description{
Deprecated functions. Will be removed in a future version.

\itemize{
\item get_mfa_quanti_var(). Deprecated. Use get_mfa_var(res.mfa, "quanti.var") instead.
\item get_mfa_quali_var().  Deprecated. Use get_mfa_var(res.mfa, "quali.var") instead.
\item get_mfa_group(). Deprecated. Use get_mfa_var(res.mfa, "group") instead.

\item{fviz_mfa_ind_starplot(): Star graph of individuals (draws partial points).
Deprecated. Use fviz_mfa_ind(res.mfa, partial = "all") instead.}
\item{fviz_mfa_quanti_var(): Graph of quantitative variables. Deprecated. Use fviz_mfa(X, "quanti.var") instead.}
\item{fviz_mfa_quali_var(): Graph of qualitative variables. Deprecated. Use fviz_mfa(X, "quali.var") instead.}

 \item
get_hmfa_quanti_var(). Deprecated. Use get_hmfa_var(res.hmfa, "quanti.var")
instead. \item get_hmfa_quali_var().  Deprecated. Use get_hmfa_var(res.hmfa,
"quali.var") instead. \item get_hmfa_group(). Deprecated. Use
get_hmfa_var(res.hmfa, "group") instead.

\item{fviz_hmfa_ind_starplot(): Graph of partial individuals. Deprecated. Use fviz_hmfa_ind(X, partial = "all") instead.}
\item{fviz_hmfa_quanti_var(): Graph of quantitative variables. Deprecated. Use fviz_hmfa_var(X, "quanti.var") instead.}
\item{fviz_hmfa_quali_var(): Graph of qualitative variables. Deprecated. Use fviz_hmfa_var(X, "quali.var") instead.}
\item{fviz_hmfa_group(): Graph of the groups representation. Deprecated. Use fviz_hmfa_var(X, "group") instead.}

}
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
