% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eigenvalue.R
\name{eigenvalue}
\alias{eigenvalue}
\alias{fviz_eig}
\alias{fviz_screeplot}
\alias{get_eig}
\alias{get_eigenvalue}
\title{Extract and visualize the eigenvalues/variances of dimensions}
\usage{
get_eig(X)

get_eigenvalue(X)

fviz_eig(X, choice = c("variance", "eigenvalue"), geom = c("bar", "line"),
  barfill = "steelblue", barcolor = "steelblue", linecolor = "black",
  ncp = 10, addlabels = FALSE, ...)

fviz_screeplot(...)
}
\arguments{
\item{X}{an object of class PCA, CA and MCA [FactoMineR]; prcomp and princomp [stats];
dudi, pca, coa and acm [ade4]; ca and mjca [ca package].}

\item{choice}{a text specifying the data to be plotted.
Allowed values are "variance" or "eigenvalue".}

\item{geom}{a text specifying the geometry to be used for the graph.
Allowed values are "bar" for barplot, "line" for lineplot or c("bar", "line") to use both types.}

\item{barfill}{fill color for bar plot.}

\item{barcolor}{outline color for bar plot.}

\item{linecolor}{color for line plot (when geom contains "line").}

\item{ncp}{a numeric value specifying the number of dimensions to be shown.}

\item{addlabels}{logical value. If TRUE, labels are added at the top of bars or points
showing the information retained by each dimension.}

\item{...}{optional arguments to be passed to the functions geom_bar(),
 geom_line(), geom_text() or fviz_eig().}
}
\value{
\itemize{
\item{get_eig() (or get_eigenvalue()): returns a data.frame containing 3 columns:
the eigenvalues, the percentage of variance and  the cumulative percentage of variance
retained by each dimension.}
\item{fviz_eig() (or fviz_screeplot()): returns a ggplot2}
}
}
\description{
Extracts and plots the eigenvalues/variances of the dimensions
from the results of Principal Component Analysis (PCA),
Correspondence Analysis (CA) and
Multiple Correspondence Analysis (MCA) functions.\cr\cr
\itemize{
\item{get_eig(): Extract the eigenvalues/variances of the principal dimensions}
\item{fviz_eig(): Plot the eigenvalues/variances against the number of dimensions}
\item{get_eigenvalue(): an alias of get_eig()}
\item{fviz_screeplot(): an alias of fviz_eig()}
}
}
\examples{
\donttest{
# Principal Component Analysis
# ++++++++++++++++++++++++++
data(iris)
res.pca <- prcomp(iris[, -5],  scale = TRUE)

# Extract eigenvalues/variances
get_eig(res.pca)

# Default plot
fviz_eig(res.pca)

# Add labels
fviz_eig(res.pca, addlabels=TRUE)

# Scree plot - Eigenvalues
fviz_eig(res.pca, choice = "eigenvalue", addlabels=TRUE)

# Use only bar plot
fviz_eig(res.pca, geom="bar", width=0.8)

# Use only line plot
fviz_eig(res.pca, geom="line")

# Change theme
fviz_screeplot(res.pca) + theme_minimal()
# theme_classic()
fviz_eig(res.pca) + theme_classic()

# Customized plot
fviz_eig(res.pca, addlabels=TRUE, hjust = -0.3,
           linecolor ="red") + theme_minimal()
# Change colors, y axis limits and theme
p <- fviz_eig(res.pca, addlabels=TRUE, hjust = -0.3,
               barfill="white", barcolor ="darkblue",
               linecolor ="red") + ylim(0, 85) +
               theme_minimal()
print(p)
# Change titles
p + labs(title = "Variances - PCA",
        x = "Principal Components", y = "\% of variances")

# Correspondence Analysis
# +++++++++++++++++++++++++++++++++
library(FactoMineR)
data(housetasks)
res.ca <- CA(housetasks, graph = FALSE)
get_eig(res.ca)
fviz_eig(res.ca)

# Multiple Correspondence Analysis
# +++++++++++++++++++++++++++++++++
library(FactoMineR)
data(poison)
res.mca <- MCA(poison, quanti.sup = 1:2,
              quali.sup = 3:4, graph=FALSE)
get_eig(res.mca)
fviz_eig(res.mca)
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

