% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_add.R
\name{fviz_add}
\alias{fviz_add}
\title{Add supplementary data to a plot}
\usage{
fviz_add(ggp, df, axes = c(1, 2), geom = c("point", "arrow"),
  color = "blue", addlabel = TRUE, labelsize = 4, shape = 19,
  linetype = "dashed")
}
\arguments{
\item{ggp}{a ggplot2 plot.}

\item{df}{a data frame containing the x and y coordinates}

\item{axes}{a numeric vector of length 2 specifying the components to be plotted.}

\item{geom}{a character specifying the geometry to be used for the graph
Allowed values are "point" or "arrow"}

\item{color}{the color to be used}

\item{addlabel}{a logical value. If TRUE, labels are added}

\item{labelsize}{the size of labels. Default value is 4}

\item{shape}{point shape when geom ="point"}

\item{linetype}{the linetype to be used when geom ="arrow"}
}
\value{
a ggplot2 plot
}
\description{
Add supplementary data to a plot
}
\examples{
\donttest{
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

