% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz_mfa.R
\name{fviz_mfa}
\alias{fviz_mfa}
\alias{fviz_mfa_axes}
\alias{fviz_mfa_group}
\alias{fviz_mfa_ind}
\alias{fviz_mfa_quali_biplot}
\alias{fviz_mfa_quali_var}
\alias{fviz_mfa_quanti_var}
\title{Visualize Multiple Factor Analysis}
\usage{
fviz_mfa_ind(X, axes = c(1, 2), geom = c("point", "text"), label = "all",
  invisible = "none", labelsize = 4, pointsize = 2, habillage = "none",
  addEllipses = FALSE, ellipse.level = 0.95, col.ind = "blue",
  col.ind.sup = "darkblue", alpha.ind = 1, shape.ind = 19,
  select.ind = list(name = NULL, cos2 = NULL, contrib = NULL),
  jitter = list(what = "label", width = NULL, height = NULL), ...)

fviz_mfa_quanti_var(X, axes = c(1, 2), geom = c("arrow", "text"),
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  col.var = "red", alpha.var = 1, shape.var = 17,
  col.quanti.sup = "blue", col.quali.sup = "darkgreen",
  col.circle = "grey70", select.var = list(name = NULL, cos2 = NULL, contrib
  = NULL), jitter = list(what = "label", width = NULL, height = NULL))

fviz_mfa_quali_var(X, axes = c(1, 2), geom = c("point", "text"),
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  col.var = "red", alpha.var = 1, shape.var = 17,
  col.quanti.sup = "blue", col.quali.sup = "darkgreen",
  select.var = list(name = NULL, cos2 = NULL, contrib = NULL),
  jitter = list(what = "label", width = NULL, height = NULL))

fviz_mfa_group(X, axes = c(1, 2), geom = c("point", "text"),
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  col.quanti.sup = "blue", col.quali.sup = "darkgreen",
  select.var = list(name = NULL, cos2 = NULL, contrib = NULL),
  jitter = list(what = "label", width = NULL, height = NULL), ...)

fviz_mfa_axes(X, axes = c(1, 2), geom = c("arrow", "text"), label = "all",
  invisible = "none", labelsize = 4, pointsize = 2, habillage = "none",
  addEllipses = FALSE, ellipse.level = 0.95, col.ind = "blue",
  col.ind.sup = "darkblue", alpha.ind = 1, col.var = "red",
  alpha.var = 1, col.quanti.sup = "blue", col.quali.sup = "darkgreen",
  col.circle = "grey70", shape.ind = 19, shape.var = 17,
  select.var = list(name = NULL, cos2 = NULL, contrib = NULL),
  select.ind = list(name = NULL, cos2 = NULL, contrib = NULL),
  arrows = c(FALSE, FALSE), jitter = list(what = "label", width = NULL,
  height = NULL), ...)

fviz_mfa_quali_biplot(X, axes = c(1, 2), geom = c("point", "text"),
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  habillage = "none", addEllipses = FALSE, ellipse.level = 0.95,
  col.ind = "blue", col.ind.sup = "darkblue", alpha.ind = 1,
  col.var = "red", alpha.var = 1, col.quanti.sup = "blue",
  col.quali.sup = "darkgreen", shape.ind = 19, shape.var = 17,
  select.var = list(name = NULL, cos2 = NULL, contrib = NULL),
  select.ind = list(name = NULL, cos2 = NULL, contrib = NULL),
  arrows = c(FALSE, FALSE), jitter = list(what = "label", width = NULL,
  height = NULL), ...)

fviz_mfa(X, ...)
}
\arguments{
\item{X}{an object of class MFA [FactoMineR].
#@inheritParams fviz_mca}

\item{label}{a text specifying the elements to be labelled.
 Default value is "all".
 Allowed values are "none" or the combination of c("ind", "ind.sup","var", "quali.sup",  "quanti.sup").
 "ind" can be used to label only active individuals.
 "ind.sup" is for supplementary individuals.
"var" is for active variable categories.
 "quali.sup" is for supplementary qualitative variable categories.
"quanti.sup" is for quantitative supplementary variables.}

\item{invisible}{a text specifying the elements to be hidden on the plot.
Default value is "none".
Allowed values are the combination of c("ind", "ind.sup","var", "quali.sup",  "quanti.sup").}

\item{habillage}{an optional factor variable for coloring
the observations by groups. Default value is "none".
If X is an MFA object from FactoMineR package, habillage can also specify
the index of the factor variable in the data.}

\item{col.ind, col.var}{color for individuals and variables, respectively.
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for individuals/variables are automatically controlled by their qualities ("cos2"),
contributions ("contrib"), coordinates (x^2 + y^2 , "coord"), x values("x") or y values("y").
To use automatic coloring (by cos2, contrib, ....), make sure that habillage ="none".}

\item{alpha.ind, alpha.var}{controls the transparency of
 individual and variable colors, respectively.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for individual/variable colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2 + y^2 , "coord"), x values("x") or y values("y").
 To use this, make sure that habillage ="none".}

\item{shape.ind, shape.var}{point shapes of individuals and variables}

\item{select.ind, select.var}{a selection of individuals/variables to be drawn.
Allowed values are NULL or a list containing the arguments name, cos2 or contrib:
\itemize{
\item name is a character vector containing individuals/variables to be drawn
\item cos2 if cos2 is in [0, 1], ex: 0.6, then individuals/variables with a cos2 > 0.6 are drawn.
if cos2 > 1, ex: 5, then the top 5 individuals/variables with the highest cos2 are drawn.
\item contrib if contrib > 1, ex: 5,  then the top 5 individuals/variables with the highest cos2 are drawn
}}

\item{jitter}{a parameter used to jitter the points in order to reduce overplotting. 
It's a list containing the objects what, width and height (i.e jitter = list(what, width, height)).
Otherwise you can use ggrepel for an elegant solution. TODO: ADD FUNCTIONALITY FOR TEXT ONLY!!! 
\itemize{
\item what: the element to be jittered. Possible values are "point" or "p"; "label" or "l"; "both" or "b"
\item width: degree of jitter in x direction
\item height: degree of jitter in y direction
}}

\item{...}{Arguments to be passed to the function fviz_mfa_quali_biplot()}

\item{col.quanti.sup, col.quali.sup}{a color for the quantitative/qualitative supplementary variables.}

\item{arrows}{Vector of two logicals specifying if the plot should contain
points (FALSE, default) or arrows (TRUE).
First value sets the rows and the second value sets the columns.}
}
\value{
a ggplot2 plot
}
\description{
Graph of individuals/quantitative variables/qualitative variables/group/partial axes from the output of Multiple Factor Analysis (MFA).\cr\cr
\itemize{
\item{fviz_mfa_ind(): Graph of individuals}
\item{fviz_mfa_quanti_var(): Graph of quantitative variables}
\item{fviz_mfa_quali_var(): Graph of qualitative variables}
\item{fviz_mfa_group(): Graph of the groups representation}
\item{fviz_mfa_axes(): Graph of partial axes}
\item{fviz_mfa_quali_biplot(): Biplot of individuals and qualitative variables}
\item{fviz_mfa(): An alias of fviz_mca_biplot()}
}
}
\details{
The default plot of MFA is a "symmetric" plot in which both rows and
columns are in principal coordinates. In this situation, it's not possible
to interpret the distance between row points and column points. To overcome this
problem, the simplest way is to make an asymmetric plot. This means that,
the column profiles must be presented in row space or vice-versa.
The allowed options for the argument map are:
\itemize{
\item "rowprincipal" or "colprincipal": asymmetric plots with either rows in principal
coordinates and columns in standard coordinates, or vice versa.
These plots preserve row metric or column metric respectively.
\item "symbiplot": Both rows and columns are scaled to have variances
equal to the singular values (square roots of eigenvalues),
which gives a symmetric biplot but does not preserve row or column metrics.
\item "rowgab" or "colgab": Asymmetric maps, proposed by Gabriel & Odoroff (1990),
with rows (respectively, columns) in
principal coordinates and columns (respectively, rows) in standard coordinates
multiplied by the mass of the corresponding point.
\item "rowgreen" or "colgreen": The so-called contribution biplots
showing visually the most contributing points (Greenacre 2006b).
These are similar to "rowgab" and "colgab" except that the points
in standard coordinates are multiplied by the square root of the corresponding masses,
giving reconstructions of the standardized residuals.
}
}
\examples{
\donttest{
# Multiple Factor Analysis
# ++++++++++++++++++++++++
# Install and load FactoMineR to compute MFA
# install.packages("FactoMineR")
library("FactoMineR")
data(poison)
res.mfa <- MFA(poison, group=c(2,2,5,6), type=c("s","n","n","n"),
               name.group=c("desc","desc2","symptom","eat"),
               num.group.sup=1:2, graph=FALSE)

# Graph of individuals
# ++++++++++++++++++++
# Default plot
fviz_mfa_ind(res.mfa)
# Change title and axis labels
fviz_mfa_ind(res.mfa) +
 labs(title = "MFA", x = "Dim.1", y ="Dim.2" )
# Change axis limits by specifying the min and max
fviz_mfa_ind(res.mfa) +
   xlim(-0.8, 1.5) + ylim (-1.5, 1.5)
# Use text only
fviz_mfa_ind(res.mfa, geom = "text")
# Use points only
fviz_mfa_ind(res.mfa, geom="point")
# Change the size of points
fviz_mfa_ind(res.mfa, geom="point", pointsize = 4)
# Change point color and theme
fviz_mfa_ind(res.mfa, col.ind = "blue")+
   theme_minimal()
# Reduce overplotting
fviz_mfa_ind(res.mfa, jitter = list(width = 0.2, height = 0.2))

# Control automatically the color of individuals
# using the cos2 or the contributions
# cos2 = the quality of the individuals on the factor map
fviz_mfa_ind(res.mfa, col.ind="cos2")
# Gradient color
fviz_mfa_ind(res.mfa, col.ind="cos2") +
      scale_color_gradient2(low="white", mid="blue",
      high="red", midpoint=0.4)
# Change the theme and use only points
fviz_mfa_ind(res.mfa, col.ind="cos2", geom = "point") +
      scale_color_gradient2(low="white", mid="blue",
      high="red", midpoint=0.4)+ theme_minimal()

# Color by the contributions
fviz_mfa_ind(res.mfa, col.ind="contrib") +
      scale_color_gradient2(low="white", mid="blue",
      high="red", midpoint=1.5)

# Control the transparency of the color by the
# contributions
fviz_mfa_ind(res.mfa, alpha.ind="contrib") +
     theme_minimal()

# Color individuals by groups
grp <- as.factor(poison.active[, "Vomiting"])
fviz_mfa_ind(res.mfa, label="none", habillage=grp)
# Add ellipses
p <- fviz_mfa_ind(res.mfa, label="none", habillage=grp,
             addEllipses=TRUE, ellipse.level=0.95)
print(p)
# Change group colors using RColorBrewer color palettes
p + scale_color_brewer(palette="Dark2") +
   theme_minimal()
p + scale_color_brewer(palette="Paired") +
     theme_minimal()
p + scale_color_brewer(palette="Set1") +
     theme_minimal()

# Select and visualize individuals with cos2 >= 0.4
fviz_mfa_ind(res.mfa, select.ind = list(cos2 = 0.4))
# Select the top 20 according to the cos2
fviz_mfa_ind(res.mfa, select.ind = list(cos2 = 20))
# Select the top 20 contributing individuals
fviz_mfa_ind(res.mfa, select.ind = list(contrib = 20))
# Select by names
fviz_mfa_ind(res.mfa,
select.ind = list(name = c("44", "38", "53",  "39")))


# Graph of variable categories
# ++++++++++++++++++++++++++++
# Default plot
fviz_mcf_var(res.mfa)
# Change color and theme
fviz_mfa_var(res.mfa, col.var="steelblue")+
 theme_minimal()

# Control variable colors using their contributions
fviz_mfa_var(res.mfa, col.var = "contrib")+
 scale_color_gradient2(low = "white", mid = "blue",
           high = "red", midpoint = 2) +
 theme_minimal()
# Control the transparency of variables using their contributions
fviz_mfa_var(res.mfa, alpha.var = "contrib") +
   theme_minimal()

# Select and visualize categories with cos2 >= 0.4
fviz_mfa_var(res.mfa, select.var = list(cos2 = 0.4))
# Select the top 10 contributing variable categories
fviz_mfa_var(res.mfa, select.var = list(contrib = 10))
# Select by names
fviz_mfa_var(res.mfa,
 select.var= list(name = c("Courg_n", "Fever_y", "Fever_n")))

# biplot
# ++++++++++++++++++++++++++
fviz_mfa_quali_biplot(res.mfa)
# Keep only the labels for variable categories
fviz_mfa_quali_biplot(res.mfa, label ="var")
# Keep only labels for individuals
fviz_mfa_quali_biplot(res.mfa, label ="ind")
# Hide variable categories
fviz_mfa_quali_biplot(res.mfa, invisible ="var")
# Hide individuals
fviz_mfa_quali_biplot(res.mfa, invisible ="ind")
# Control automatically the color of individuals using the cos2
fviz_mfa_quali_biplot(res.mfa, label ="var", col.ind="cos2") +
       theme_minimal()
# Change the color by groups, add ellipses
fviz_mfa_quali_biplot(res.mfa, label="var", col.var ="blue",
   habillage=grp, addEllipses=TRUE, ellipse.level=0.95) +
   theme_minimal()

# Select the top 30 contributing individuals
# And the top 10 variables
fviz_mfa_quali_biplot(res.mfa,
               select.ind = list(contrib = 30),
               select.var = list(contrib = 10))

 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}

Fabian Mundt \email{f.mundt@inventionate.de}
}
\references{
http://www.sthda.com
}

