% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_pca_contrib.R
\name{fviz_pca_contrib}
\alias{fviz_pca_contrib}
\title{Contributions of variables/individuals - Principal Component Analysis}
\usage{
fviz_pca_contrib(X, choice = c("var", "ind"), axes = 1,
  fill = "steelblue", color = "steelblue", sortcontrib = c("desc", "asc",
  "none"), top = Inf, ...)
}
\arguments{
\item{X}{an object of class PCA (FactoMineR); prcomp (stats); princomp (stats);
dudi and pca (ade4).}

\item{choice}{allowed values are "var" (for variable contributions) or "ind"
for the contribution of individuals}

\item{axes}{a numeric vector specifying the component(s) of interest.}

\item{fill}{a fill color for the bar plot}

\item{color}{an outline color the bar plot}

\item{sortcontrib}{a string specifying whether the contributions should be sorted.
Allowed values are "none" (no sorting), "asc" (for ascending) or "desc" (for descending)}

\item{top}{a numeric value specifing the top contributing elements to be shown}

\item{...}{optional arguments to be passed to the function get_pca_ind().
 This can includes the argument data (the original data used for pca)
 which is required when X is not from FactoMineR or adea4 packages.}
}
\value{
a ggplot2 plot
}
\description{
This function can be used to visualize the contributions of variables/individuals from the output of several PCA functions :
PCA() from FactoMineR package; prcomp() and princomp() from stats package;
 dudi.pca() from ade4 package.
}
\examples{
\donttest{
data(decathlon2)
decathlon2.active <- decathlon2[1:23, 1:10]
# Principal component analysis
res.pca <- princomp(decathlon2.active,  cor = TRUE)

# variable contributions on axis 1
fviz_pca_contrib(res.pca, choice="var", axes = 1 )
# sorting
fviz_pca_contrib(res.pca, choice="var", axes = 1, sort = "asc" )

variable contributions on axis 2
fviz_pca_contrib(res.pca, choice="var", axes = 2, sort = "asc" )

# Contributions of individuals on axis 1
fviz_pca_contrib(res.pca, choice="var", axes = 1, sort = "asc" )
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

