% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_pca_var.R
\name{fviz_pca_var}
\alias{fviz_pca_var}
\title{Visualizing Principal Component Analysis Variables using ggplot2}
\usage{
fviz_pca_var(X, axes = c(1, 2), label = "all", invisible = "none",
  labelsize = 4, col.var = "black", alpha.var = 1,
  col.quanti.sup = "blue", col.circle = "grey70", autolab = FALSE)
}
\arguments{
\item{X}{an object of class PCA (from FactoMineR package).}

\item{axes}{a numeric vector of length 2 specifying the component to be plotted.}

\item{label}{a character vector specifying the elements to be labelled.
Default value is "all". Allowed values are "none" or the combination of c("var", "quanti.sup").
"var" can be used to label only active variables.
"quanti.sup" can be used to label only quantitative supplementary variables.}

\item{invisible}{a character value specifying the elements to be hidden on the plot.
Default value is "none". Allowed values are the combination of c("var", "quanti.sup").}

\item{labelsize}{font size for the labels}

\item{col.var}{color for variables. The default value is "black".
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for variables are automatically controlled by their qualities ("cos2"),
contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y")}

\item{alpha.var}{controls the transparency of colors.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for variable colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").}

\item{col.quanti.sup}{a color for the quantitative supplementary variables.}

\item{col.circle}{a color for the correlation circle.}

\item{autolab}{logical value. If TRUE, the best positions for labels are calculated to avoid overlapping.
 This can be time-consuming if there are many variables.}
}
\value{
a ggplot2 plot
}
\description{
Visualizing Principal Component Analysis Variables using ggplot2
}
\examples{
\donttest{
 library("FactoMineR")
 data(decathlon)
 res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup=13, graph = FALSE)
 fviz_pca_var(res.pca, col.var="contrib")

 # Change the color
 p <- fviz_pca_var(res.pca, col.var="contrib")
 p + scale_color_gradient2(low ="blue", mid="white", high="red", midpoint=40)

 # Change the theme
 p + theme_minimal()

 p + theme_classic()
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

