% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_eigenvalue.R
\name{get_eigenvalue}
\alias{get_eigenvalue}
\title{Extract eigenvalues/variances - Principal Component Analysis}
\usage{
get_eigenvalue(X)
}
\arguments{
\item{X}{an object of class PCA (FactoMineR); prcomp (stats); princomp (stats);
 dudi and pca (ade4).}
}
\value{
a data.frame containing 3 columns :
 eigenvalue, percentage of variance and cumulative percentage of variance.
}
\description{
Ectract eigenvalues/variances from the results of several PCA functions :
PCA() from FactoMineR package; prcomp() and princomp() from stats package;
 dudi.pca() from ade4 package.
}
\examples{
\donttest{
 res.pca <- princomp(iris[, -5],  cor = TRUE)
 eig <- get_eigenvalue(res.pca)
 head(eig)
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

