% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca.R
\name{get_pca}
\alias{get_pca}
\alias{get_pca_ind}
\alias{get_pca_var}
\title{Extract the results for individuals/variables - PCA}
\usage{
get_pca(res.pca, element = c("var", "ind"))

get_pca_ind(res.pca, ...)

get_pca_var(res.pca)
}
\arguments{
\item{res.pca}{an object of class PCA [FactoMineR]; 
\code{prcomp} or \code{princomp} [stats]; \code{factoextra_pca}; \code{pca},
\code{dudi}, \code{between}, or \code{within} [ade4]; or
\code{expoOutput}/\code{epPCA} [ExPosition].}

\item{element}{the element to subset from the output. Allowed values are 
"var" (for active variables) or "ind" (for active individuals).}

\item{...}{not used}
}
\value{
a list of matrices containing all the results for the active individuals/variables including: 
\item{coord}{coordinates for the individuals/variables}
\item{cos2}{cos2 for the individuals/variables}
\item{contrib}{contributions of the individuals/variables; contributions to
each nonzero-inertia axis sum to 100 percent, while a zero-inertia axis
contains zeros}
\item{cor}{loading-times-component-standard-deviation coordinates for PCA
objects from \code{stats}; these equal variable-component correlations when
the input variables were standardized. Returned for variables.}
}
\description{
Extract all the results (coordinates, squared cosines, and contributions) for
the active individuals/variables from Principal Component Analysis (PCA) outputs.\cr\cr
\itemize{
\item get_pca(): Extract the results for variables and individuals
\item get_pca_ind(): Extract the results for individuals only
\item get_pca_var(): Extract the results for variables only
}
}
\examples{
\donttest{
# Principal Component Analysis
# +++++++++++++++++++++++++++++
 data(iris)
 res.pca <- prcomp(iris[, -5],  scale = TRUE)
 # Extract the results for individuals
 ind <- get_pca_ind(res.pca)
 print(ind)
 head(ind$coord) # coordinates of individuals
 head(ind$cos2) # cos2 of individuals
 head(ind$contrib) # contributions of individuals
 
 # Extract the results for variables
 var <- get_pca_var(res.pca)
 print(var)
 head(var$coord) # coordinates of variables
 head(var$cos2) # cos2 of variables
 head(var$contrib) # contributions of variables
 
 # You can also use the function get_pca()
 get_pca(res.pca, "ind") # Results for individuals
 get_pca(res.pca, "var") # Results for variable categories
 }
}
\references{
\url{https://www.sthda.com/english/}
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
