% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_pca_ind.R
\name{get_pca_ind}
\alias{get_pca_ind}
\title{Extract the results for the individuals after a principal component analysis}
\usage{
get_pca_ind(res.pca, data = NULL)
}
\arguments{
\item{res.pca}{an object of class PCA (from FactoMineR);
prcomp and princomp (from stats package); pca, dudi (dudi.pca, from adea4 package).}

\item{data}{the original data used for the pca.
This argument is required only when res.pca is not from FactoMineR.
It's used to calculate the cos2 of the individuals.}
}
\value{
a list of matrices containing all the results for the active individuals including :
\item{coord}{coordinates for the individuals}
\item{cos2}{cos2 for the individuals}
\item{contrib}{contributions of the individuals}
}
\description{
Ectract all the results for the active individuals from a principal component analysis output.
 This output contains individuals coordinates, square cosine and contributions.
 The allowed PCA outputs are the ones from FactoMineR (PCA),
 stats (princomp() and prcomp()), ade4 (dudi.pca()) packages.
}
\examples{
\donttest{
 res.pca <- princomp(iris[, -5])
 ind <- get_pca_ind(res.pca)

 head(ind$coord)

 head(ind$cos2)

 head(ind$contrib)
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

