% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_pca_var.R
\name{get_pca_var}
\alias{get_pca_var}
\title{Extract the results for the variables after a principal component analysis}
\usage{
get_pca_var(res.pca)
}
\arguments{
\item{res.pca}{an object of class PCA (from FactoMineR);
prcomp and princomp (from stats package); pca, dudi (dudi.pca, from adea4 package)}
}
\value{
a list of matrices containing all the results for the active variables including :
\item{coord}{coordinates for the variables}
\item{cor}{correlations between variables and dimensions}
\item{cos2}{cos2 for the variables}
\item{contrib}{contributions of the variables}
}
\description{
Ectract all the results for the active variables from a principal component analysis output.
 This output contains variable coordinates,
 correlation between variables and axes, square cosine and contributions.
 The allowed PCA outputs are the ones from FactoMineR (PCA),
 stats (princomp() and prcomp()), ade4 (dudi.pca()) packages.
}
\examples{
\donttest{
 res.pca <- iris[, -5]
 var <- get_pca_var(res.pca)
 var
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

