% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.factoextra.R
\name{print.factoextra}
\alias{print.factoextra}
\title{Print method for an object of class factoextra}
\usage{
\method{print}{factoextra}(x, ...)
}
\arguments{
\item{x}{an object of class factoextra}

\item{...}{further arguments to be passed to print method}
}
\description{
Print method for an object of class factoextra
}
\examples{
\donttest{
 data(iris)
 res.pca <- princomp(iris[, -5],  cor = TRUE)
 ind <- get_pca_ind(res.pca, data = iris[, -5])
 print(ind)
 }
 
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

