% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz_pca.R
\name{fviz_pca}
\alias{fviz_pca}
\alias{fviz_pca_biplot}
\alias{fviz_pca_ind}
\alias{fviz_pca_var}
\title{Visualize Principal Component Analysis}
\usage{
fviz_pca(X, ...)

fviz_pca_ind(X, axes = c(1, 2), geom = c("point", "text"), repel = FALSE,
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  habillage = "none", addEllipses = FALSE, ellipse.level = 0.95,
  col.ind = "black", col.ind.sup = "blue", alpha.ind = 1,
  select.ind = list(name = NULL, cos2 = NULL, contrib = NULL),
  jitter = list(what = "label", width = NULL, height = NULL), ...)

fviz_pca_var(X, axes = c(1, 2), geom = c("arrow", "text"), label = "all",
  invisible = "none", repel = FALSE, labelsize = 4, col.var = "black",
  alpha.var = 1, col.quanti.sup = "blue", col.circle = "grey70",
  select.var = list(name = NULL, cos2 = NULL, contrib = NULL),
  jitter = list(what = "label", width = NULL, height = NULL))

fviz_pca_biplot(X, axes = c(1, 2), geom = c("point", "text"),
  label = "all", invisible = "none", labelsize = 4, pointsize = 2,
  habillage = "none", addEllipses = FALSE, ellipse.level = 0.95,
  col.ind = "black", col.ind.sup = "blue", alpha.ind = 1,
  col.var = "steelblue", alpha.var = 1, col.quanti.sup = "blue",
  col.circle = "grey70", repel = FALSE, select.var = list(name = NULL,
  cos2 = NULL, contrib = NULL), select.ind = list(name = NULL, cos2 = NULL,
  contrib = NULL), jitter = list(what = "label", width = NULL, height = NULL),
  ...)
}
\arguments{
\item{X}{an object of class PCA [FactoMineR]; prcomp and princomp [stats];
dudi and pca [ade4].}

\item{...}{Arguments to be passed to the function fviz_pca_biplot().}

\item{axes}{a numeric vector of length 2 specifying the dimensions to be plotted.}

\item{geom}{a text specifying the geometry to be used for the graph. 
Allowed values are the combination of c("point", "arrow", "text"). 
Use "point" (to show only points); 
"text" to show only labels; c("point", "text") or c("arrow", "text") to show both types.}

\item{repel}{a boolean, whether to use ggrepel to avoid overplotting text labels or not.}

\item{label}{a text specifying the elements to be labelled.
Default value is "all".
Allowed values are "none" or the combination of c("ind", "ind.sup", "quali", "var", "quanti.sup").
"ind" can be used to label only active individuals.
"ind.sup" is for supplementary individuals.
"quali" is for supplementary qualitative variables. "var" is for active variables.
"quanti.sup" is for quantitative supplementary variables.}

\item{invisible}{a text specifying the elements to be hidden on the plot.
Default value is "none".
Allowed values are the combination of c("ind", "ind.sup", "quali", "var", "quanti.sup").}

\item{labelsize}{font size for the labels}

\item{pointsize}{the size of points}

\item{habillage}{an optional factor variable for coloring
the observations by groups. Default value is "none".
If X is a PCA object from FactoMineR package, habillage can also specify
the supplementary qualitative variable (by its index or name) to be used for
coloring individuals by groups (see ?PCA in FactoMineR).}

\item{addEllipses}{logical value.
If TRUE, draws ellipses around the individuals when habillage != "none".}

\item{ellipse.level}{the size of the concentration ellipse in normal probability}

\item{col.ind, col.var}{color for individuals and variables, respectively.
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for individuals/variables are automatically controlled by their 
qualities of representation ("cos2"),
contributions ("contrib"), coordinates (x^2+y^2, "coord"), x values ("x") or y values ("y").
To use automatic coloring (by cos2, contrib, ....), make sure that habillage ="none".}

\item{col.ind.sup}{color for supplementary individuals}

\item{alpha.ind, alpha.var}{controls the transparency of
 individual and variable colors, respectively.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for the individual/variable colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2+y^2, "coord"), x values("x") or y values("y").
 To use this, make sure that habillage ="none".}

\item{select.ind, select.var}{a selection of individuals/variables to be drawn. 
Allowed values are NULL or a list containing the arguments name, cos2 or contrib: 
\itemize{
\item name: is a character vector containing individuals/variables to be drawn
\item cos2: if cos2 is in [0, 1], ex: 0.6, then individuals/variables with a cos2 > 0.6 are drawn. 
if cos2 > 1, ex: 5, then the top 5 individuals/variables with the highest cos2 are drawn.
\item contrib: if contrib > 1, ex: 5,  then the top 5 individuals/variables with the highest contrib are drawn
}}

\item{jitter}{a parameter used to jitter the points in order to reduce overplotting. 
It's a list containing the objects what, width and height (i.e jitter = list(what, width, height)). 
\itemize{
\item what: the element to be jittered. Possible values are "point" or "p"; "label" or "l"; "both" or "b".
\item width: degree of jitter in x direction
\item height: degree of jitter in y direction
}}

\item{col.quanti.sup}{a color for the quantitative supplementary variables.}

\item{col.circle}{a color for the correlation circle.}
}
\value{
a ggplot2 plot
}
\description{
Graph of individuals/variables from the output of Principal Component Analysis (PCA). \cr\cr
\itemize{
\item{fviz_pca_ind(): Graph of individuals}
\item{fviz_pca_var(): Graph of variables}
\item{fviz_pca_biplot(): Biplot of individuals and variables}
\item{fviz_pca(): An alias of fviz_pca_biplot()}
}
}
\examples{
\donttest{
# Principal component analysis
# ++++++++++++++++++++++++++++++
data(iris)
res.pca <- prcomp(iris[, -5],  scale = TRUE)

# Graph of individuals
# +++++++++++++++++++++
# Default plot
fviz_pca_ind(res.pca)
# Change title and axis labels
fviz_pca_ind(res.pca) +
 labs(title = "PCA", x = "PC1", y ="PC2" )
# Change axis limits by specifying the min and max
fviz_pca_ind(res.pca) + 
   xlim(-4, 4) + ylim (-4, 4)
# Use text only
fviz_pca_ind(res.pca, geom="text")
# Use points only
fviz_pca_ind(res.pca, geom="point")
# Change the size of points
fviz_pca_ind(res.pca, geom="point", pointsize = 4)
# Change point color and theme
fviz_pca_ind(res.pca, col.ind = "blue")+
   theme_minimal()
   
# Control automatically the color of individuals 
# using the cos2 or the contributions
# cos2 = the quality of the individuals on the factor map
fviz_pca_ind(res.pca, col.ind="cos2") 
# Gradient color
fviz_pca_ind(res.pca, col.ind="cos2") + 
      scale_color_gradient2(low="white", mid="blue", 
      high="red", midpoint=0.6, space = "Lab")
# Change the theme and use only points
fviz_pca_ind(res.pca, col.ind="cos2", geom = "point") + 
      scale_color_gradient2(low="blue", mid="white", 
      high="red", midpoint=0.6, space = "Lab")+ theme_minimal()
      
# Color by the contributions   
fviz_pca_ind(res.pca, col.ind="contrib") + 
      scale_color_gradient2(low="blue", mid="white", 
      high="red", midpoint=4, space = "Lab")
      
# Control the transparency of the color by the
# contributions
fviz_pca_ind(res.pca, alpha.ind="contrib") +
     theme_minimal()        
             
# Color individuals by groups
fviz_pca_ind(res.pca, label="none", habillage=iris$Species)
# Add ellipses
p <- fviz_pca_ind(res.pca, label="none", habillage=iris$Species, 
             addEllipses=TRUE, ellipse.level=0.95)
print(p)
             
# Change group color using RColorBrewer color palettes
p + scale_color_brewer(palette="Dark2") +
     theme_minimal()
p + scale_color_brewer(palette="Paired") +
     theme_minimal()
p + scale_color_brewer(palette="Set1") +
     theme_minimal()
# Change color manually
p + scale_color_manual(values=c("#999999", "#E69F00", "#56B4E9"))           
# Select and visualize individuals with cos2 >= 0.96
fviz_pca_ind(res.pca, select.ind = list(cos2 = 0.96))
# Select the top 20 according to the cos2
fviz_pca_ind(res.pca, select.ind = list(cos2 = 20))
# Select the top 20 contributing individuals
fviz_pca_ind(res.pca, select.ind = list(contrib = 20))
# Select by names
fviz_pca_ind(res.pca, 
select.ind = list(name = c("23", "42", "119")))

 
# Graph of variables
# ++++++++++++++++++++++++++++
# Default plot
fviz_pca_var(res.pca)
# Use points and text
fviz_pca_var(res.pca, geom = c("point", "text"))
# Change color and theme
fviz_pca_var(res.pca, col.var="steelblue")+
 theme_minimal()
 
# Control variable colors using their contributions
fviz_pca_var(res.pca, col.var="contrib")+
 scale_color_gradient2(low="white", mid="blue", 
           high="red", midpoint=96, space = "Lab") +
 theme_minimal()          
# Control the transparency of variables using their contributions
fviz_pca_var(res.pca, alpha.var="contrib") +
   theme_minimal()
   
# Select and visualize variables with cos2 >= 0.96
fviz_pca_var(res.pca, select.var = list(cos2 = 0.96))
# Select the top 3 contributing variables
fviz_pca_var(res.pca, select.var = list(contrib = 3))
# Select by names
fviz_pca_var(res.pca, 
   select.var= list(name = c("Sepal.Width", "Petal.Length")))
    
# biplot
# ++++++++++++++++++++++++++
fviz_pca_biplot(res.pca)
# Keep only the labels for variables
fviz_pca_biplot(res.pca, label ="var")
# Keep only labels for individuals
fviz_pca_biplot(res.pca, label ="ind")
# Hide variables
fviz_pca_biplot(res.pca, invisible ="var")
# Hide individuals
fviz_pca_biplot(res.pca, invisible ="ind")
# Control automatically the color of individuals using the cos2
fviz_pca_biplot(res.pca, label ="var", col.ind="cos2") +
       theme_minimal()
# Change the color by groups, add ellipses
fviz_pca_biplot(res.pca, label="var", habillage=iris$Species,
               addEllipses=TRUE, ellipse.level=0.95) 
               
# Select the top 30 contributing individuals
fviz_pca_biplot(res.pca, label="var", 
               select.ind = list(contrib = 30)) 

 }
 
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

