% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_pca_biplot.R
\name{fviz_pca_biplot}
\alias{fviz_pca_biplot}
\title{Biplot of Principal Component Analysis Individuals and Variables using ggplot2}
\usage{
fviz_pca_biplot(X, axes = c(1, 2), geom = c("point", "text"),
  label = "all", invisible = "none", labelsize = 4, habillage = "none",
  addEllipses = FALSE, ellipse.level = 0.95, col.ind = "black",
  col.ind.sup = "blue", alpha.ind = 1, col.var = "steelblue",
  alpha.var = 1, col.quanti.sup = "blue", col.circle = "grey70",
  data = NULL, ...)
}
\arguments{
\item{X}{an object of class PCA (from FactoMineR package).}

\item{axes}{a numeric vector of length 2 specifying the component to be plotted.}

\item{geom}{a character specifying the geometry to be used for the graph.
Allowed values are "point" (to show only points),
"text" to show only labels or c("point", "text") to show both types.}

\item{label}{a character vector specifying the elements to be labelled.
Default value is "all".
Allowed values are "none" or the combination of c("ind", "ind.sup", "quali", "var", "quanti.sup").
"ind" can be used to label only active individuals.
"ind.sup" if for supplementary individuals.
"quali" is for supplementary qualitative variables. "var" is for active variables.
"quanti.sup" is for quantitative supplementary variables.}

\item{invisible}{a character value specifying the elements to be hidden on the plot.
Default value is "none".
Allowed values are the combination of c("ind", "ind.sup", "quali", "var", "quanti.sup").}

\item{labelsize}{font size for the labels}

\item{habillage}{specify the qualitative variable (by its index or name) to be used for
coloring individuals by groups. Default value is "none".}

\item{addEllipses}{logical value.
If TRUE, draw ellipses around the individuals when habillage != "none".}

\item{ellipse.level}{the size of the concentration ellipse in normal probability}

\item{col.ind,}{col.var color for individuals and variables, respectively.
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for individuals/variables are automatically controlled by their qualities ("cos2"),
contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").
To use automatic coloring (by cos2, contrib, ....), make sure that habillage ="none".}

\item{col.ind.sup}{color for supplementary individuals}

\item{alpha.ind,}{alpha.var controls the transparency of
 individual and variable colors, respectively.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for individuals/variables colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").
 To use this, make sure that habillage ="none".}

\item{col.quanti.sup}{a color for the quantitative supplementary variables.}

\item{col.circle}{a color for the correlation circle.}

\item{data}{the original data used for the pca.
This argument is required only when X is not from FactoMineR.
It's used to calculate the cos2 of the individuals.}

\item{...}{optional arguments to be passed to the function get_pca_ind().
This can includes the argument data (the original data used for pca)
which is required when X is not from FactoMineR.}
}
\value{
a ggplot2 plot
}
\description{
Biplot of Principal Component Analysis Individuals and Variables using ggplot2
}
\examples{
\donttest{
 library("FactoMineR")
 data(decathlon)
 res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup=13, graph = FALSE)
 fviz_pca_biplot(res.pca)

 # Use habillage
 p <- fviz_pca_biplot(res.pca, habillage=13, label=c("var", "quanti.sup"))
 p

 # Change the theme
 p + theme_minimal()

 p + theme_classic()
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

