% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_pca_ind.R
\name{fviz_pca_ind}
\alias{fviz_pca_ind}
\title{Graph of individuals - Principal Component Analysis}
\usage{
fviz_pca_ind(X, axes = c(1, 2), geom = c("point", "text"), label = "all",
  invisible = "none", labelsize = 4, habillage = "none",
  addEllipses = FALSE, ellipse.level = 0.95, col.ind = "black",
  col.ind.sup = "blue", alpha.ind = 1, ...)
}
\arguments{
\item{X}{an object of class PCA (FactoMineR); prcomp (stats); princomp (stats);
dudi and pca (ade4).}

\item{axes}{a numeric vector of length 2 specifying the components to be plotted.}

\item{geom}{a character specifying the geometry to be used for the graph.
Allowed values are "point" (to show only points),
"text" to show only labels or c("point", "text") to show both types.}

\item{label}{a character vector specifying the elements to be labelled.
Default value is "all".
Allowed values are "none" or the combination of c("ind", "ind.sup", "quali").
"ind" can be used to label only active individuals.
"ind.sup" is for supplementary individuals.
"quali" is for supplementary qualitative variables.}

\item{invisible}{a character value specifying the elements to be hidden on the plot.
Default value is "none". Allowed values are the combination of c("ind", "ind.sup", "quali").}

\item{labelsize}{font size for the labels}

\item{habillage}{an optional factor variable for coloring
the observations by groups. Default value is "none".
If X is an PCA object from FactoMineR package, habillage can also specify
the supplementary qualitative variable (by its index or name) to be used for
coloring individuals by groups (see ?PCA in FactoMineR).}

\item{addEllipses}{logical value.
If TRUE, draws ellipses around the individuals when habillage != "none".}

\item{ellipse.level}{the size of the concentration ellipse in normal probability}

\item{col.ind}{color for individuals. The default value is "black".
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for individuals are automatically controlled by their qualities ("cos2"),
contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").
To use automatic coloring (by cos2, contrib, ....), make sure that habillage ="none".}

\item{col.ind.sup}{color for supplementary individuals}

\item{alpha.ind}{controls the transparency of colors.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for individuals colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").
 To use this, make sure that habillage ="none".}

\item{...}{optional arguments to be passed to the function get_pca_ind().
This can includes the argument data (the original data used for pca)
which is required when X is not from FactoMineR or adea4 packages.}
}
\value{
a ggplot2 plot
}
\description{
This function can be used to visualize individual factor maps from the output of several PCA functions :
PCA() from FactoMineR package; prcomp() and princomp() from stats package;
 dudi.pca() from ade4 package.
}
\examples{
\donttest{
data(iris)
# Principal component analysis
res.pca <- princomp(iris[, -5],  cor = TRUE)

# Default plot
fviz_pca_ind(res.pca, data = iris[, -5])

# Remove the points from the graph
fviz_pca_ind(res.pca, data = iris[, -5], geom="text")

 # Color individuals by groups
 fviz_pca_ind(res.pca, data = iris[, -5], label="none",
 habillage=iris$Species, addEllipses=TRUE, ellipse.level=0.95)
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

