% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_pca_var.R
\name{fviz_pca_var}
\alias{fviz_pca_var}
\title{Graph of variables - Principal Component Analysis}
\usage{
fviz_pca_var(X, axes = c(1, 2), label = "all", invisible = "none",
  labelsize = 4, col.var = "black", alpha.var = 1,
  col.quanti.sup = "blue", col.circle = "grey70")
}
\arguments{
\item{X}{an object of class PCA (FactoMineR); prcomp (stats); princomp (stats);
dudi and pca (ade4).}

\item{axes}{a numeric vector of length 2 specifying the component to be plotted.}

\item{label}{a character vector specifying the elements to be labelled.
Default value is "all". Allowed values are "none" or the combination of c("var", "quanti.sup").
"var" can be used to label only active variables.
"quanti.sup" can be used to label only quantitative supplementary variables.}

\item{invisible}{a character value specifying the elements to be hidden on the plot.
Default value is "none". Allowed values are the combination of c("var", "quanti.sup").}

\item{labelsize}{font size for the labels}

\item{col.var}{color for variables. The default value is "black".
Possible values include also : "cos2", "contrib", "coord", "x" or "y".
In this case, the colors for variables are automatically controlled by their qualities ("cos2"),
contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y")}

\item{alpha.var}{controls the transparency of colors.
The value can variate from 0 (total transparency) to 1 (no transparency).
Default value is 1. Possible values include also : "cos2", "contrib", "coord", "x" or "y".
 In this case, the transparency for variable colors are automatically controlled by their qualities ("cos2"),
 contributions ("contrib"), coordinates (x^2+y2, "coord"), x values("x") or y values("y").}

\item{col.quanti.sup}{a color for the quantitative supplementary variables.}

\item{col.circle}{a color for the correlation circle.}
}
\value{
a ggplot2 plot
}
\description{
This function can be used to visualize variable factor maps from the output of several PCA functions :
PCA() from FactoMineR package; prcomp() and princomp() from stats package;
 dudi.pca() from ade4 package.
}
\examples{
\donttest{
data(iris)
# Principal component analysis
res.pca <- princomp(iris[, -5],  cor = TRUE)
# Default plot
fviz_pca_var(res.pca)
# Change color and theme
fviz_pca_var(res.pca, col.var="steelblue")+
 theme_minimal()
# Control variable colors using their contribution
fviz_pca_var(res.pca, col.var="contrib")
# Change the gradient color
fviz_pca_var(res.pca, col.var="contrib")+
 scale_color_gradient2(low="blue", mid="white",
           high="red", midpoint=55)+theme_bw()
# Control the transparency of variables using their contribution
fviz_pca_var(res.pca, alpha.var="contrib")+
   theme_minimal()
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

