% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fviz_screeplot.R
\name{fviz_screeplot}
\alias{fviz_screeplot}
\title{Plot the variances/eigenvalues against the number of dimensions}
\usage{
fviz_screeplot(X, choice = c("variance", "eigenvalue"), geom = c("bar",
  "line"), barfill = "steelblue", barcolor = "steelblue",
  linecolor = "black", ncp = 5, addlabels = FALSE, ...)
}
\arguments{
\item{X}{an object of class PCA (FactoMineR); prcomp (stats); princomp (stats);
dudi and pca (ade4)}

\item{choice}{a text specifyinng the type of the data to be plotted.
Allowed values are "variance" or "eigenvalue"}

\item{geom}{a text specifying the geometry to be used for the graph.
Allowed values are "bar" for barplot, "line" for lineplot or c("bar", "line") to use both type}

\item{barfill}{fill color for bar plot}

\item{barcolor}{outline color for bar plot}

\item{linecolor}{color for line plot (when geom contains "line")}

\item{ncp}{a numeric value specifying the number of components to be shown}

\item{addlabels}{logical value. If TRUE, labels are added at the top of bars or points
 showing the information retained by each dimension}
}
\value{
a ggplot2 plot
}
\description{
Plot the variances/eigenvalues against the number of dimensions
}
\examples{
\donttest{
data(iris)
res.pca <- princomp(iris[, -5],  cor = TRUE)

# Scree plot
fviz_screeplot(res.pca, addlabels=TRUE)

# Use only barplot
fviz_screeplot(res.pca, geom="bar", width=0.8)

# Change theme
fviz_screeplot(res.pca) + theme_minimal()
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

