% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clust_tendency.R
\name{get_clust_tendency}
\alias{get_clust_tendency}
\title{Assessing Clustering Tendency}
\usage{
get_clust_tendency(data, n, graph = TRUE, gradient = list(low = "red", mid =
  "white", high = "blue"), seed = 123)
}
\arguments{
\item{data}{a numeric data frame or matrix. Columns are variables and rows are samples. 
Computation are done on rows (samples) by default. If you want to calculate Hopkins statistic on 
variables, transpose the data before.}

\item{n}{the number of points selected from sample space which is also the number of 
points selected from the given sample(data).}

\item{graph}{logical value; if TRUE the ordered dissimilarity image (ODI) is shown.}

\item{gradient}{a list containing three elements specifying the colors for low, mid and high values in 
the ordered dissimilarity image. The element "mid" can take the value of NULL.}

\item{seed}{an integer specifying the seed for random number generator. Specify seed for reproducible results.}
}
\value{
A list containing the elements: 
       
- hopkins_stat for Hopkins statistic value
      
- plot for ordered dissimilarity image. This is generated using the function fviz_dist(dist.obj).
}
\description{
Assess clustering tendency using Hopkins' statistic and visual approach. 
An ordered dissimilarity image (ODI) is shown. Objects belonging to the same cluster are displayed in consecutive order 
using hierarchical clustering.
}
\examples{
data(iris)

# Clustering tendency
get_clust_tendency(iris[,-5], 50)

# Customize the dissimilarity image
fviz_dist(dist(iris[, -5]), 
   gradient = list(low = "black", mid = NULL, high = "white"))
}

