% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_clust_tendency.R
\name{get_clust_tendency}
\alias{get_clust_tendency}
\title{Assessing clustering tendency}
\usage{
get_clust_tendency(data, n, graph = TRUE, gradient = list(low = "black",
  high = "white"), seed = 123)
}
\arguments{
\item{data}{a numeric data frame or matrix. Columns are variables and rows are samples.
Computation are done on rows (samples) by default. If you want to calculate Hopkins statistic on
variables, transpose the data before.}

\item{n}{the number of points selected from sample space which is also the number of
points selected from the given sample(data).}

\item{graph}{logical value; if TRUE the ordered dissimilarity image (ODI) is shown}

\item{gradient}{a list containing two element specifying the colors for low and high values in
the ordered dissimilarity image.}

\item{seed}{an integer specifying the seed for random number generator. Specify seed for reproducible results.}
}
\value{
A list containing the element "hopkins_stat" for Hopkins statistic value
}
\description{
Assess clustering tendency using Hopkins' statistic and visual approach.
An ordered dissimilarity image (ODI) is shown. Objects belonging to the same cluster are displayed in consecutive order.
}
\examples{
\donttest{
data(iris)
get_clust_tendency(iris[,-5], 50)
}
}

