% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_eigenvalue.R
\name{get_eigenvalue}
\alias{get_eigenvalue}
\title{Extract eigenvalues/variances after factor analysis}
\usage{
get_eigenvalue(X)
}
\arguments{
\item{X}{Object of class PCA, MCA and CA (from FactoMineR);
 prcomp and princomp (from stats package); pca, dudi (dudi.pca, from adea4 package)}
}
\value{
a data.frame containing 3 columns :
 eigenvalue, percentage of variance and cumulative percentage of variance.
}
\description{
Ectract eigenvalues/variances from the results of different PCA, MCA and CA functions in R.
Results from FactoMineR (PCA, MCA, CA functions), ade4 (dudi.pca function)
and stats (prcomp, princomp functions) packages are allowed.
}
\examples{
\donttest{
 res.pca <- princomp(iris[, -5],  cor = TRUE)
 eig <- get_eigenvalue(res.pca)
 head(eig)
 }
}
\author{
Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com}
}
\references{
http://www.sthda.com
}

