% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison.R
\docType{data}
\name{poison}
\alias{poison}
\title{Poison}
\format{
A data frame with 55 rows and 15 columns.
}
\source{
This data is from FactoMineR package.
}
\usage{
data("poison")
}
\description{
This data is a result from a survey carried out on children of
 primary school who suffered from food poisoning. They were asked about their
 symptoms and about what they ate.
}
\examples{
\donttest{
library(FactoMineR)
data(poison)
res.mca <- MCA(poison, quanti.sup = 1:2, quali.sup = c(3,4), 
   graph = FALSE)
fviz_mca_biplot(res.mca, repel = TRUE)+
theme_minimal()

}

}
